Make the atom radius (used for compact atom elements) a double rather than an int.
git-svn-id: https://bioclipse.svn.sourceforge.net/svnroot/bioclipse/bioclipse2/trunk@8831 fcb5ba71-d80d-0410-8237-ba3920747fcc

+++ cdk-externals/trunk/org.openscience.cdk.render/src/org/openscience/cdk/renderer/RendererModel.java
@@ -523,19 +523,19 @@ public class RendererModel implements Serializable, Cloneable {
     }
 
     /**
-     * XXX No idea what this is about.
+     * This is used for the size of the compact atom element. 
      */
-    public int getAtomRadius() {
+    public double getAtomRadius() {
         return this.parameters.getAtomRadius();
     }
 
     /**
-     * XXX No idea what this is about.
+     * Set the radius of the compact atom representation.
      * 
-     * @param atomRadius
-     *            XXX No idea what this is about
+     * @param atomRadius the size of the compact atom symbol.
+     *
      */
-    public void setAtomRadius(int atomRadius) {
+    public void setAtomRadius(double atomRadius) {
         this.parameters.setAtomRadius(atomRadius);
         fireChange();
     }
